trying to fix bugs
