+++ Project.toml
@@ -22,7 +22,7 @@ DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 
 [extensions]
-DiffEqExt = ["DifferentialEquations", "OrdinaryDiffEq", "DelayDiffEq"]
+AlgebraicDynamicsDiffEqExt = ["DifferentialEquations", "OrdinaryDiffEq", "DelayDiffEq"]
 
 
 [compat]
